+++ src/gmap.ml
@@ -26,36 +26,39 @@ module type S = sig
   val mem : 'a key -> t -> bool
   val find : 'a key -> t -> 'a option
   val get : 'a key -> t -> 'a
+  val add_unbound : 'a key -> 'a -> t -> t option
   val add : 'a key -> 'a -> t -> t
   val remove : 'a key -> t -> t
+  val update : 'a key -> ('a option -> 'a option) -> t -> t
 
-  type v = V : 'a key * 'a -> v
-  val min_binding : t -> v option
-  val max_binding : t -> v option
-  val any_binding :  t -> v option
-  val bindings : t -> v list
+  type b = B : 'a key * 'a -> b
+  val min_binding : t -> b option
+  val max_binding : t -> b option
+  val any_binding :  t -> b option
+  val bindings : t -> b list
 
-  val findv : 'a key -> t -> v option
-  val getv : 'a key -> t -> v
+  val findb : 'a key -> t -> b option
+  val getb : 'a key -> t -> b
 
-  val addv : v -> t -> t
+  val addb_unbound : b -> t -> t option
+  val addb : b -> t -> t
 
-  val equal : (v -> v -> bool) -> t -> t -> bool
+  val equal : (b -> b -> bool) -> t -> t -> bool
 
-  val iter : (v -> unit) -> t -> unit
-  val fold : (v -> 'a -> 'a) -> t -> 'a -> 'a
-  val for_all : (v -> bool) -> t -> bool
-  val exists : (v -> bool) -> t -> bool
-  val filter : (v -> bool) -> t -> t
-  val merge : (v option -> v option -> v option) -> t -> t -> t
-  val union : (v -> v -> v option) -> t -> t -> t
+  val iter : (b -> unit) -> t -> unit
+  val fold : (b -> 'a -> 'a) -> t -> 'a -> 'a
+  val for_all : (b -> bool) -> t -> bool
+  val exists : (b -> bool) -> t -> bool
+  val filter : (b -> bool) -> t -> t
+  val merge : (b option -> b option -> b option) -> t -> t -> t
+  val union : (b -> b -> b option) -> t -> t -> t
   val pp : Format.formatter -> t -> unit
 end
 
 module Make (Key : KEY) : S with type 'a key = 'a Key.t = struct
   type 'a key = 'a Key.t
   type k = K : 'a key -> k
-  type v = V : 'a key * 'a -> v
+  type b = B : 'a key * 'a -> b
 
   module M = Map.Make(struct
       type t = k
@@ -65,62 +68,72 @@ module Make (Key : KEY) : S with type 'a key = 'a Key.t = struct
         | Order.Gt -> 1
     end)
 
-  type t = v M.t
+  type t = b M.t
 
   let empty = M.empty
-  let singleton k v = M.singleton (K k) (V (k, v))
+  let singleton k v = M.singleton (K k) (B (k, v))
 
   let is_empty = M.is_empty
   let mem k m = M.mem (K k) m
 
-  let addv (V (k, _) as v) m = M.add (K k) v m
-  let add k v m = M.add (K k) (V (k, v)) m
+  let add k v m = M.add (K k) (B (k, v)) m
+  let addb (B (k, v)) m = add k v m
+
+  let add_unbound k v m = if mem k m then None else Some (add k v m)
+
+  let addb_unbound (B (k, v)) m = add_unbound k v m
 
   let remove k m = M.remove (K k) m
 
-  let getv : type a. a Key.t -> t -> v = fun k t ->
-    match M.find (K k) t with
-    | V (k', v) ->
+  let getb : type a. a key -> t -> b = fun k m ->
+    match M.find (K k) m with
+    | B (k', v) ->
       match Key.compare k k' with
-      | Order.Eq -> V (k, v)
+      | Order.Eq -> B (k, v)
       | _ -> assert false
-  let get : type a. a Key.t -> t -> a = fun k t ->
-    match M.find (K k) t with
-    | V (k', v) ->
+
+  let get : type a. a key -> t -> a = fun k m ->
+    match M.find (K k) m with
+    | B (k', v) ->
       match Key.compare k k' with
       | Order.Eq -> v
       | _ -> assert false
 
-  let findv : type a. a Key.t -> t -> v option = fun k t ->
-    try Some (getv k t) with Not_found -> None
-  let find : type a. a Key.t -> t -> a option = fun k t ->
-    try Some (get k t) with Not_found -> None
+  let findb : type a. a key -> t -> b option = fun k m ->
+    try Some (getb k m) with Not_found -> None
 
-  let any_binding t = try Some (snd (M.choose t)) with Not_found -> None
-  let min_binding t = try Some (snd (M.min_binding t)) with Not_found -> None
-  let max_binding t = try Some (snd (M.max_binding t)) with Not_found -> None
-  let bindings t = snd (List.split (M.bindings t))
+  let find : type a. a key -> t -> a option = fun k m ->
+    try Some (get k m) with Not_found -> None
 
-  let cardinal t = M.cardinal t
+  let update k f m =
+    match f (find k m) with
+    | None -> remove k m
+    | Some v -> add k v m
 
-  let for_all f t = M.for_all (fun _ b -> f b) t
-  let exists f t = M.exists (fun _ b -> f b) t
 
-  let iter f t = M.iter (fun _ b -> f b) t
-  let fold f t acc = M.fold (fun _ b acc -> f b acc) t acc
-  let filter f t = M.filter (fun _ b -> f b) t
+  let any_binding m = try Some (snd (M.choose m)) with Not_found -> None
+  let min_binding m = try Some (snd (M.min_binding m)) with Not_found -> None
+  let max_binding m = try Some (snd (M.max_binding m)) with Not_found -> None
+  let bindings m = snd (List.split (M.bindings m))
 
-  let merge f a b =
-    M.merge (fun _ v1 v2 -> f v1 v2) a b
+  let cardinal m = M.cardinal m
 
-  let union f a b =
-    M.union (fun _ v1 v2 -> f v1 v2) a b
+  let for_all p m = M.for_all (fun _ b -> p b) m
+  let exists p m = M.exists (fun _ b -> p b) m
 
-  let pp ppf t =
+  let iter f m = M.iter (fun _ b -> f b) m
+  let fold f m acc = M.fold (fun _ b acc -> f b acc) m acc
+  let filter p m = M.filter (fun _ b -> p b) m
+
+  let merge f m m' = M.merge (fun _ b b' -> f b b') m m'
+
+  let union f m m' = M.union (fun _ b b' -> f b b') m m'
+
+  let pp ppf m =
     let pp ppf = function
-      | V (k, v) -> Key.pp ppf k v
+      | B (k, v) -> Key.pp ppf k v
     in
-    Fmt.(list ~sep:(unit "@.") pp) ppf (bindings t)
+    Fmt.(list ~sep:(unit "@.") pp) ppf (bindings m)
 
-  let equal cmp a b = M.equal cmp a b
+  let equal cmp m m' = M.equal cmp m m'
 end