+++ CHANGES.md
@@ -1,3 +1,9 @@
+## master
+
+* New function `update`.
+* New function `add_unbound` and `addb_unbound`.
+* Replace `type v = V : 'a key * 'a -> v` by `type b = B : 'a key * 'a -> b`.
+
 ## 0.1.0 (2018-06-16)
 
 * Initial release